feat(discoveryTableSearchHighlighting): cleaned up code some
